comit sales

+++ src/services/cart.service.js
@@ -1,9 +1,9 @@
 /* eslint-disable no-await-in-loop */
+const mongoose = require('mongoose');
 const httpStatus = require('http-status');
-const { Cart, CartItem, ProductCatalogue, Sales, Ledger, SalesItem } = require('../models');
+const { Cart, CartItem, ProductCatalogue, Ledger, Sales, SalesItem } = require('../models');
 const ApiError = require('../utils/ApiError');
-const { ACCOUNT_TYPE, DIRECTION_VALUE } = require('../constants/account');
-const { SALES_PAYMENT_METHOD } = require('../constants/sales');
+const { ACCOUNT_TYPE, DIRECTION_VALUE, SALES_PAYMENT_METHOD } = require('../constants/account');
 
 /**
  * Initialize a new cart for the user
@@ -29,28 +29,32 @@ const initCart = async (userId, total) => {
  * @returns {Promise<Object>} The updated cart and added cart item
  */
 const addToCart = async (userId, productCatalogueId, unitPrice, quantity) => {
-  let cart = await Cart.findOne({ userId });
-  if (!cart) {
-    // Initialize Cart
+  try {
+    let cart = await Cart.findOne({ userId });
+    if (!cart) {
+      // Initialize Cart
+      const subTotal = unitPrice * quantity;
+      cart = await initCart(userId, subTotal);
+    }
+
     const subTotal = unitPrice * quantity;
-    cart = await initCart(userId, subTotal);
+    const cartId = cart._id;
+
+    const addItem = await CartItem.create({
+      cartId,
+      productCatalogueId,
+      unitPrice,
+      quantity,
+      subTotal,
+    });
+
+    cart.total += subTotal;
+    await cart.save();
+
+    return { cart, addItem };
+  } catch (error) {
+    throw new ApiError(httpStatus.INTERNAL_SERVER_ERROR, 'Error adding item to cart');
   }
-
-  const subTotal = unitPrice * quantity;
-  const cartId = cart._id;
-
-  const addItem = await CartItem.create({
-    cartId,
-    productCatalogueId,
-    unitPrice,
-    quantity,
-    subTotal,
-  });
-
-  cart.total += subTotal;
-  await cart.save();
-
-  return { cart, addItem };
 };
 
 /**
@@ -59,12 +63,16 @@ const addToCart = async (userId, productCatalogueId, unitPrice, quantity) => {
  * @returns {Promise<Object>} The cart and its items
  */
 const getCartItems = async (userId) => {
-  const cart = await Cart.findOne({ userId });
-  if (!cart) {
-    throw new ApiError(httpStatus.NOT_FOUND, 'Cart item not found');
+  try {
+    const cart = await Cart.findOne({ userId });
+    if (!cart) {
+      throw new ApiError(httpStatus.NOT_FOUND, 'Cart item not found');
+    }
+    const cartItems = await CartItem.find({ cartId: cart._id });
+    return { cart, cartItems };
+  } catch (error) {
+    throw new ApiError(httpStatus.INTERNAL_SERVER_ERROR, 'Error retrieving cart items');
   }
-  const cartItems = await CartItem.find({ cartId: cart._id });
-  return { cart, cartItems };
 };
 
 /**
@@ -74,18 +82,22 @@ const getCartItems = async (userId) => {
  * @returns {Promise<Object>} The updated cart and removed cart item
  */
 const removeCartItem = async (userId, productCatalogueId) => {
-  const cart = await Cart.findOne({ userId });
-  if (!cart) {
-    throw new ApiError(httpStatus.NOT_FOUND, 'Cart item not found');
+  try {
+    const cart = await Cart.findOne({ userId });
+    if (!cart) {
+      throw new ApiError(httpStatus.NOT_FOUND, 'Cart item not found');
+    }
+    const cartItem = await CartItem.findOneAndRemove({
+      productCatalogueId,
+      cartId: cart._id,
+    });
+    const subTotal = cartItem ? +cartItem.subTotal : 0;
+    cart.total -= parseFloat(subTotal);
+    const result = await cart.save();
+    return result;
+  } catch (error) {
+    throw new ApiError(httpStatus.INTERNAL_SERVER_ERROR, 'Error removing cart item');
   }
-  const cartItem = await CartItem.findOneAndRemove({
-    productCatalogueId,
-    cartId: cart._id,
-  });
-  const subTotal = cartItem ? +cartItem.subTotal : 0;
-  cart.total -= parseFloat(subTotal);
-  const result = await cart.save();
-  return result;
 };
 
 /**
@@ -94,10 +106,17 @@ const removeCartItem = async (userId, productCatalogueId) => {
  * @returns {Promise<Object>} The deleted cart
  */
 const clearCart = async (userId) => {
-  const cart = await Cart.findOne({ userId });
-  await CartItem.deleteMany({ cartId: cart._id });
-  const deletedCart = await cart.remove();
-  return deletedCart;
+  try {
+    const cart = await Cart.findOne({ userId });
+    if (!cart) {
+      throw new ApiError(httpStatus.NOT_FOUND, 'Cart item not found');
+    }
+    await CartItem.deleteMany({ cartId: cart._id });
+    const deletedCart = await cart.remove();
+    return deletedCart;
+  } catch (error) {
+    throw new ApiError(httpStatus.INTERNAL_SERVER_ERROR, 'Error clearing cart');
+  }
 };
 
 const checkProductAvailability = async (productCatalogueId, quantityDemand) => {
@@ -109,116 +128,94 @@ const checkProductAvailability = async (productCatalogueId, quantityDemand) => {
   return false;
 };
 
+/**
+ * Update inventory
+ * @param {ObjectId} productCatalogueId
+ * @param {Object} quantityDemand
+ * @returns {Promise<User>}
+ */
 const updateInventory = async (productCatalogueId, quantityDemand, session) => {
-  try {
-    const productCatalogue = await ProductCatalogue.findById(productCatalogueId);
-    if (!productCatalogue) {
-      throw new ApiError(httpStatus.NOT_FOUND, 'Product not found');
-    }
-
-    const newQuantity = productCatalogue.quantity - quantityDemand;
-    await ProductCatalogue.findByIdAndUpdate(productCatalogueId, { quantity: newQuantity }, { session });
-
-    return { message: 'success', error: null };
-  } catch (err) {
-    return { message: null, error: err.message };
+  const productCatalogue = await ProductCatalogue.findById(productCatalogueId);
+  if (!productCatalogue) {
+    throw new ApiError(httpStatus.NOT_FOUND, 'Product not found');
   }
+  const newQuantity = productCatalogue.quantity - quantityDemand;
+  productCatalogue.quantity = newQuantity;
+  await ProductCatalogue.findByIdAndUpdate(productCatalogueId, { quantity: newQuantity }, { session });
 };
 
 /**
  * Commit a sale by creating sales records, updating inventory, and posting ledger entries.
- * @param {string} userId - The ID of the user
- * @param {Object} salesData - The sales data
- * @param {Object} session - The database session
+ * @param {string} userId - The ID of the user initiating the sale.
+ * @param {Object} salesData - The sales data.
+ * @param {Object} session - The database session object for transactional consistency.
  * @returns {Promise<Object>} The result of the sale operation
  */
-const commitSale = async (userId, salesData, session) => {
+const commitSale = async (userId, salesData) => {
+  const session = await mongoose.startSession();
+  session.startTransaction();
+
   try {
-    // Retrieve the user's cart
-    const cart = await Cart.findOne({ userId });
-    const cartId = cart._id;
-    const { total } = cart;
-    // Retrieve cart items
-    const cartItems = await CartItem.find({ cartId });
-    if (!cartItems.length) {
-      throw new ApiError(httpStatus.NOT_FOUND, 'No Item in the Cart');
+    const cart = await Cart.findOne({ userId }).session(session);
+    if (!cart) {
+      throw new ApiError(httpStatus.NOT_FOUND, 'Cart not found');
     }
 
-    // Create a sales record
-    const createSale = await Sales.create(
-      [
-        {
-          ...salesData,
-          userId,
-          total,
-        },
-      ],
-      { session }
-    );
-    console.log(createSale[0]);
-    // Process each cart item
+    const cartId = cart._id;
+    const { total } = cart;
+
+    const cartItems = await CartItem.find({ cartId }).session(session);
+    if (cartItems.length === 0) {
+      throw new ApiError(httpStatus.BAD_REQUEST, 'No items in the cart');
+    }
+
+    const sales = await Sales.create([{ ...salesData, userId, total, branchId: salesData.branchId }], { session });
+
     for (let i = 0; i < cartItems.length; i += 1) {
       const item = cartItems[i];
       const isAvailable = await checkProductAvailability(item.productCatalogueId, item.quantity);
       if (isAvailable) {
-        // Create a sales item record
-        try {
-          console.log(item.subTotal, item.quantity, item.unitPrice, item.productCatalogueId, createSale[0]._id);
-          const saleRecord = await SalesItem.create(
+        await SalesItem.create(
+          [
             {
-              subTotal: item.subTotal,
+              salesId: sales[0]._id,
+              productCatalogueId: item.productCatalogueId,
               quantity: item.quantity,
               unitPrice: item.unitPrice,
-              productCatalogueId: item.productCatalogueId,
-              salesId: createSale[0]._id,
+              subTotal: item.subTotal,
             },
-            { session }
-          );
-          console.log(saleRecord);
-        } catch (error) {
-          console.error(error.message);
-        }
+          ],
+          { session }
+        );
 
-        // Update inventory
-        const result = await updateInventory(item.productCatalogueId, item.quantity, session);
-        if (!result) {
-          throw new ApiError(httpStatus.NOT_FOUND, 'Product Catalogue could not be updated');
+        const updateInventoryResult = await updateInventory(item.productCatalogueId, item.quantity, session);
+        if (!updateInventoryResult) {
+          throw new ApiError(httpStatus.INTERNAL_SERVER_ERROR, 'Product Catalogue could not be updated');
         }
       } else {
-        throw new ApiError(httpStatus.NOT_FOUND, 'Product not available');
+        throw new ApiError(httpStatus.BAD_REQUEST, 'Product not available');
       }
     }
 
-    if (salesData.paymentMethod === SALES_PAYMENT_METHOD[1]) {
-      // TODO: Check if the wallet balance is enough for the transaction
-      // TODO: Withdraw money from wallet before posting the ledger
-
-      const date = new Date().getTime();
+    if (SALES_PAYMENT_METHOD.includes(salesData.paymentMethod)) {
+      const currentDate = new Date().getTime();
       const amount = total;
       const type = ACCOUNT_TYPE[0];
       const direction = DIRECTION_VALUE[0];
-      const naration = salesData.salesRepId;
+      const narration = 'Sale Transaction';
       const { branchId } = salesData;
 
-      // Create a ledger entry
-      await Ledger.create(
-        {
-          date,
-          amount,
-          type,
-          direction,
-          naration,
-          branchId,
-          userId,
-        },
-        { session }
-      );
-
-      return { message: 'success', error: null };
+      await Ledger.create([{ date: currentDate, amount, type, direction, narration, branchId, userId }], { session });
     }
-    // TODO: Check for Cash or Transfer
-  } catch (err) {
-    return { message: null, error: err.message };
+
+    await session.commitTransaction();
+    session.endSession();
+
+    return { message: 'Sale committed successfully', error: null };
+  } catch (error) {
+    await session.abortTransaction();
+    session.endSession();
+    throw error;
   }
 };
